feat: Extend support of messaging to application nav
