+++ module/Olcs/src/Controller/Messages/AbstractCreateConversationController.php
@@ -16,10 +16,10 @@ use Olcs\Controller\Interfaces\LicenceControllerInterface;
 use Olcs\Form\Model\Form\Conversation;
 use RuntimeException;
 
-class AbstractCreateConversationController extends AbstractInternalController implements LeftViewProvider, LicenceControllerInterface, ToggleAwareInterface
+class AbstractCreateConversationController
+    extends AbstractInternalController
+    implements LeftViewProvider, LicenceControllerInterface, ToggleAwareInterface
 {
-    protected $navigationId = 'licence_new_conversation';
-
     protected $mapperClass = DefaultMapper::class;
 
     protected $createCommand = Create::class;
@@ -36,20 +36,9 @@ class AbstractCreateConversationController extends AbstractInternalController im
         'addAction' => ['forms/message-categories']
     ];
 
-    protected $redirectConfig = [
-        'add' => [
-            'route' => 'licence/conversation/view',
-            'resultIdMap' => [
-                'conversation' => 'conversation',
-                'licence' => 'licence'
-            ],
-            'reUseParams' => true
-        ]
-    ];
-
     public function getLeftView(): ViewModel
     {
-        $view = new ViewModel();
+        $view = new ViewModel(['navigationId' => $this->navigationId]);
         $view->setTemplate('sections/messages/partials/left');
 
         return $view;
