+++ module/Olcs/src/Controller/Messages/ApplicationConversationListController.php
@@ -26,7 +26,7 @@ class ApplicationConversationListController extends AbstractInternalController i
      */
     public function getLeftView()
     {
-        $view = new ViewModel();
+        $view = new ViewModel(['navigationId' => $this->navigationId]);
         $view->setTemplate('sections/messages/partials/left');
 
         return $view;
