+++ module/Olcs/src/Controller/Messages/LicenceCreateConversationController.php
@@ -1,10 +1,21 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Olcs\Controller\Messages;
 
-use Olcs\Controller\Interfaces\LeftViewProvider;
-
-class LicenceCreateConversationController extends AbstractCreateConversationController implements LeftViewProvider
+class LicenceCreateConversationController extends AbstractCreateConversationController
 {
-    protected $navigationId = 'licence_new_conversation';
+    protected $navigationId = 'conversations';
+
+    protected $redirectConfig = [
+        'add' => [
+            'route' => 'licence/conversation/view',
+            'resultIdMap' => [
+                'conversation' => 'conversation',
+                'licence' => 'licence'
+            ],
+            'reUseParams' => true
+        ]
+    ];
 }
