+++ module/Olcs/src/Controller/Messages/LicenceEnableDisableMessagingController.php
@@ -112,7 +112,7 @@ class LicenceEnableDisableMessagingController extends ApplicationController
             return null;
         }
 
-        $view = new ViewModel();
+        $view = new ViewModel(['navigationId' => $this->navigationId]);
         $view->setTemplate('sections/messages/partials/left');
 
         return $view;
