+++ module/Olcs/view/sections/messages/partials/left.phtml
@@ -1,5 +1,5 @@
 <h4 class="title"><?php echo $this->translate('internal-licence-messages-sidebar'); ?></h4>
 <?php
-$currentSection = $this->navigation('navigation')->findById('conversations');
+$currentSection = $this->navigation('navigation')->findById($this->navigationId);
 
 echo $this->navigation($currentSection)->menu()->setPartial('partials/vertical-message-navigation');
